CMenu::ProcessMouse: decompile scrollbar drag/click and line hover handling
Address 0x50C8A0, 1463 bytes. Handles mouse input for a CMenu: dragging
the scrollbar thumb, clicking the up/down arrows and page areas, and
hovering or clicking a text line.

Also fixes field_34 and field_38 on CMenu from i32 to f32. The original
disassembly does float ops (fld/fadd/fcomp/fstp) directly on that
memory, they hold the scrollbar drag step and accumulated drag delta.

Added gsub_4015B0, a real empty function in the original binary
(nullsub_1 at 0x4015b0) that ProcessMouse calls once per scrollbar hit
with a debug string. The string content and addresses come straight
from IDA on the original exe.

This is a functional decompile, not a byte match (see the matching bar
override for this session). Not hooked into the game, so no runtime
test.

+++ front.h
@@ -120,8 +120,14 @@ public:
 	i32 mLineSep;
 	i16 field_30;
 	u16 field_32;
-	i32 field_34;
-	i32 field_38;
+
+	// scrollbar thumb drag state, set up in CMenu::ProcessMouse's thumb-grab
+	// case (0x50c8a0): field_34 is the pixel step per visible line (drag speed),
+	// field_38 is the accumulated mouse-Y delta since the last line change.
+	// confirmed float (not int) by the fld/fadd/fcomp/fstp float ops on this
+	// memory in the original disassembly.
+	f32 field_34;
+	f32 field_38;
 
 	SEntry mEntry[40];
 };
